example-functions/demo/pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Paul Fidika", email = "paul@fidika.com" }
]
requires-python = ">=3.12"
dependencies = [
    "gen-worker",
    "Pillow>=10.0",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "mydev"
url = "http://137.184.153.104:3141/root/mydev/+simple"
explicit = true

[project.scripts]
demo = "demo:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
